# Add pools filtering