--- a/src/scripts/init.ts
+++ b/src/scripts/init.ts
@@ -1,4 +1,7 @@
-import { createPoolsTable, createTokensTable } from "../data-providers/dynamodb";
+import {
+  createPoolsTable,
+  createTokensTable,
+} from "../data-providers/dynamodb";
 import { localAWSConfig } from "../utils";
 
 const AWS = require("aws-sdk");