This pr adds possibility to pools table query with filter by `totalLiquidity > 100`