@timjrobinson 2022-09-28
I did some tests of Dev vs production for common SOR queries and the performance boost is huge! 

BAL -> DAI from 8s to 400ms
USDC -> DAI from 3s to 300ms
WETH -> BAL from 2s to 300ms

```
Testing BALDAI on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 287.175ms
Testing BALDAI on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 327.06ms
Testing BALDAI on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 306.584ms
Testing BALDAI on URL: https://api.balancer.fi/sor/1
runTest: 8.593s
Testing BALDAI on URL: https://api.balancer.fi/sor/1
runTest: 8.610s
Testing BALDAI on URL: https://api.balancer.fi/sor/1
runTest: 8.446s
Testing USDCDAI on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 361.289ms
Testing USDCDAI on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 307.085ms
Testing USDCDAI on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 258.18ms
Testing USDCDAI on URL: https://api.balancer.fi/sor/1
runTest: 2.812s
Testing USDCDAI on URL: https://api.balancer.fi/sor/1
runTest: 2.766s
Testing USDCDAI on URL: https://api.balancer.fi/sor/1
runTest: 3.891s
Testing WETHBAL on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 304.948ms
Testing WETHBAL on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 307.402ms
Testing WETHBAL on URL: https://ayuyqhk7mh.execute-api.ap-southeast-2.amazonaws.com/prod/sor/1
runTest: 307.595ms
Testing WETHBAL on URL: https://api.balancer.fi/sor/1
runTest: 1.947s
Testing WETHBAL on URL: https://api.balancer.fi/sor/1
runTest: 1.888s
Testing WETHBAL on URL: https://api.balancer.fi/sor/1
runTest: 1.080s
```